First pass on the GUI design and implementation
This initial commit is the first attempt at a GUI layout and
implementation. The GUI components are almost guaranteed to be replaced
farther down the line, but are necessary for the function of the program
and so have been made as functional pieces, if not nicely-usable pieces.
There are a handful of underlying gameplay pieces put in as well (such
as the player-character class, and the world class, but nothing works
with eachother just yet.