World view & PlayerCharacter reading & writing
The panel that shows the game view during the actual gameplay has been
implemented and work has begun on a thredworker to load worlds from file
(or socket, if there is world transfer from the server).  The class to
read and write PlayerCharacters from and to file has been created and
given basic reading abilities to populate the PC-list on the main menu.
This must be expanded so it may read all of the characters' attributes
and recreate a proper PlayerCharacter class, as well as take the PC
instance and write the attributes to disk correctly.